General Code CleanUp
Details:
- Chnaged the name of the forms from "form1, form2,. ..) to Names relative to their functionality
- Changed the names of the controls in the Main_Form from (Button1, label2, etc) to names relative to their functionality
- removed empty/un-used Events in the Main_Form
- Changed event names in Main_Form to co-respond to the new names of their respective controls
