+++ Personal Manager/bin/Debug/Personal Manager.exe.config
@@ -2,5 +2,5 @@
 <configuration>
     <startup> 
         
-    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.0"/></startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.8"/></startup>
 </configuration>
